+++ changelog.md
@@ -488,7 +488,7 @@ are 9001 for HSQLDB, 9092 for H2 and 1527 for Derby.
 - Use compact arrays for the graph and heap
 - Faster to vehicles conversion
 - Zoom-dependent simplification for the Mercator and Equal Earth projections
-
+- Faster DBF export